provault - added some useful comments, and TODO lines.
