+++ smart-contracts/contracts/pro-vault/src/vault/mod.rs
@@ -2,3 +2,4 @@ pub mod config;
 pub mod provault;
 pub mod query;
 pub mod error;
+pub mod ownership;
